+++ layoutlm/deprecated/examples/seq_labeling/run_seq_labeling.py
@@ -17,59 +17,23 @@
 
 from __future__ import absolute_import, division, print_function
 
-import argparse
-import glob
-import json
+
 import logging
 import os
 import random
 import shutil
-
 import numpy as np
 import torch
-from eval import convert_predictions_to_dict
-from seqeval.metrics import (
-    classification_report,
-    f1_score,
-    precision_score,
-    recall_score,
-)
-from pathlib import Path
-from tensorboardX import SummaryWriter
+from layoutlm import FunsdDataset
 from torch.nn import CrossEntropyLoss
-from torch.utils.data import DataLoader, RandomSampler, SequentialSampler
-from torch.utils.data.distributed import DistributedSampler
-from tqdm import tqdm, trange
-from transformers import (
-    WEIGHTS_NAME,
-    AdamW,
-    BertConfig,
-    BertForTokenClassification,
-    BertTokenizer,
-    RobertaConfig,
-    RobertaForTokenClassification,
-    RobertaTokenizer,
-    get_linear_schedule_with_warmup,
-)
 
-from layoutlm import FunsdDataset, LayoutlmConfig, LayoutlmForTokenClassification
-import torch.nn.functional as F
+from transformers import HfArgumentParser
+from train import do_train
+from predict import do_predict
+from eval import do_eval
+from training_args import TrainingArgs, MODEL_CLASSES
 
 logger = logging.getLogger(__name__)
-
-ALL_MODELS = sum(
-    (
-        tuple(conf.pretrained_config_archive_map.keys())
-        for conf in (BertConfig, RobertaConfig, LayoutlmConfig)
-    ),
-    (),
-)
-
-MODEL_CLASSES = {
-    "bert": (BertConfig, BertForTokenClassification, BertTokenizer),
-    "roberta": (RobertaConfig, RobertaForTokenClassification, RobertaTokenizer),
-    "layoutlm": (LayoutlmConfig, LayoutlmForTokenClassification, BertTokenizer),
-}
 
 
 def set_seed(args):
@@ -96,606 +60,10 @@ def get_labels(path):
     return labels
 
 
-def train(  # noqa C901
-        args, train_dataset, model, tokenizer, labels, pad_token_label_id
-):
-    """ Train the model """
-    if args.local_rank in [-1, 0]:
-        tb_writer = SummaryWriter(logdir="runs/" + os.path.basename(args.output_dir))
-
-    args.train_batch_size = args.per_gpu_train_batch_size * max(1, args.n_gpu)
-    train_sampler = (
-        RandomSampler(train_dataset)
-        if args.local_rank == -1
-        else DistributedSampler(train_dataset)
-    )
-    train_dataloader = DataLoader(
-        train_dataset,
-        sampler=train_sampler,
-        batch_size=args.train_batch_size,
-        collate_fn=None,
-    )
-
-    if args.max_steps > 0:
-        t_total = args.max_steps
-        args.num_train_epochs = (
-                args.max_steps
-                // (len(train_dataloader) // args.gradient_accumulation_steps)
-                + 1
-        )
-    else:
-        t_total = (
-                len(train_dataloader)
-                // args.gradient_accumulation_steps
-                * args.num_train_epochs
-        )
-
-    # Prepare optimizer and schedule (linear warmup and decay)
-    no_decay = ["bias", "LayerNorm.weight"]
-    optimizer_grouped_parameters = [
-        {
-            "params": [
-                p
-                for n, p in model.named_parameters()
-                if not any(nd in n for nd in no_decay)
-            ],
-            "weight_decay": args.weight_decay,
-        },
-        {
-            "params": [
-                p
-                for n, p in model.named_parameters()
-                if any(nd in n for nd in no_decay)
-            ],
-            "weight_decay": 0.0,
-        },
-    ]
-    optimizer = AdamW(
-        optimizer_grouped_parameters, lr=args.learning_rate, eps=args.adam_epsilon
-    )
-    scheduler = get_linear_schedule_with_warmup(
-        optimizer, num_warmup_steps=args.warmup_steps, num_training_steps=t_total
-    )
-    if args.fp16:
-        try:
-            from apex import amp
-        except ImportError:
-            raise ImportError(
-                "Please install apex from https://www.github.com/nvidia/apex to use fp16 training."
-            )
-        model, optimizer = amp.initialize(
-            model, optimizer, opt_level=args.fp16_opt_level
-        )
-
-    # multi-gpu training (should be after apex fp16 initialization)
-    if args.n_gpu > 1:
-        model = torch.nn.DataParallel(model)
-
-    # Distributed training (should be after apex fp16 initialization)
-    if args.local_rank != -1:
-        model = torch.nn.parallel.DistributedDataParallel(
-            model,
-            device_ids=[args.local_rank],
-            output_device=args.local_rank,
-            find_unused_parameters=True,
-        )
-
-    # Train!
-    logger.info("***** Running training *****")
-    logger.info("  Num examples = %d", len(train_dataset))
-    logger.info("  Num Epochs = %d", args.num_train_epochs)
-    logger.info(
-        "  Instantaneous batch size per GPU = %d", args.per_gpu_train_batch_size
-    )
-    logger.info(
-        "  Total train batch size (w. parallel, distributed & accumulation) = %d",
-        args.train_batch_size
-        * args.gradient_accumulation_steps
-        * (torch.distributed.get_world_size() if args.local_rank != -1 else 1),
-    )
-    logger.info("  Gradient Accumulation steps = %d", args.gradient_accumulation_steps)
-    logger.info("  Total optimization steps = %d", t_total)
-
-    global_step = 0
-    tr_loss, logging_loss = 0.0, 0.0
-    model.zero_grad()
-    train_iterator = trange(
-        int(args.num_train_epochs), desc="Epoch", disable=args.local_rank not in [-1, 0]
-    )
-    set_seed(args)  # Added here for reproductibility (even between python 2 and 3)
-    for _ in train_iterator:
-        epoch_iterator = tqdm(
-            train_dataloader, desc="Iteration", disable=args.local_rank not in [-1, 0]
-        )
-        for step, batch in enumerate(epoch_iterator):
-            model.train()
-            input_ids = batch[0]
-
-            inputs = {
-                "input_ids": input_ids.to(args.device),
-                "attention_mask": batch[1].to(args.device),
-                "labels": batch[3].to(args.device),
-            }
-            if args.model_type in ["layoutlm"]:
-                inputs["bbox"] = batch[4].to(args.device)
-            inputs["token_type_ids"] = (
-                batch[2].to(args.device) if args.model_type in ["bert", "layoutlm"] else None
-            )  # RoBERTa don"t use segment_ids
-
-            outputs = model(**inputs)
-            # model outputs are always tuple in pytorch-transformers (see doc)
-            loss = outputs[0]
-
-            if args.n_gpu > 1:
-                loss = loss.mean()  # mean() to average on multi-gpu parallel training
-            if args.gradient_accumulation_steps > 1:
-                loss = loss / args.gradient_accumulation_steps
-
-            if args.fp16:
-                with amp.scale_loss(loss, optimizer) as scaled_loss:
-                    scaled_loss.backward()
-            else:
-                loss.backward()
-
-            tr_loss += loss.item()
-            if (step + 1) % args.gradient_accumulation_steps == 0:
-                if args.fp16:
-                    torch.nn.utils.clip_grad_norm_(
-                        amp.master_params(optimizer), args.max_grad_norm
-                    )
-                else:
-                    torch.nn.utils.clip_grad_norm_(
-                        model.parameters(), args.max_grad_norm
-                    )
-                optimizer.step()
-                scheduler.step()  # Update learning rate schedule
-                model.zero_grad()
-                global_step += 1
-
-                if (
-                        args.local_rank in [-1, 0]
-                        and args.logging_steps > 0
-                        and global_step % args.logging_steps == 0
-                ):
-                    # Log metrics
-                    if (
-                            args.local_rank in [-1, 0] and args.evaluate_during_training
-                    ):  # Only evaluate when single GPU otherwise metrics may not average well
-                        results, _ = evaluate(
-                            args,
-                            model,
-                            tokenizer,
-                            labels,
-                            pad_token_label_id,
-                            mode="dev",
-                        )
-                        for key, value in results.items():
-                            tb_writer.add_scalar(
-                                "eval_{}".format(key), value, global_step
-                            )
-                    tb_writer.add_scalar("lr", scheduler.get_lr()[0], global_step)
-                    tb_writer.add_scalar(
-                        "loss",
-                        (tr_loss - logging_loss) / args.logging_steps,
-                        global_step,
-                    )
-                    logging_loss = tr_loss
-
-                if (
-                        args.local_rank in [-1, 0]
-                        and args.save_steps > 0
-                        and global_step % args.save_steps == 0
-                ):
-                    # Save model checkpoint
-                    output_dir = os.path.join(
-                        args.output_dir, "checkpoint-{}".format(global_step)
-                    )
-                    if not os.path.exists(output_dir):
-                        os.makedirs(output_dir)
-                    model_to_save = (
-                        model.module if hasattr(model, "module") else model
-                    )  # Take care of distributed/parallel training
-                    model_to_save.save_pretrained(output_dir)
-                    tokenizer.save_pretrained(output_dir)
-                    torch.save(args, os.path.join(output_dir, "training_args.bin"))
-                    logger.info("Saving model checkpoint to %s", output_dir)
-
-            if args.max_steps > 0 and global_step > args.max_steps:
-                epoch_iterator.close()
-                break
-        if args.max_steps > 0 and global_step > args.max_steps:
-            train_iterator.close()
-            break
-
-    if args.local_rank in [-1, 0]:
-        tb_writer.close()
-
-    return global_step, tr_loss / global_step
-
-
-def evaluate(args, model, tokenizer, labels, pad_token_label_id, mode, prefix=""):
-    eval_dataset = FunsdDataset(args, tokenizer, labels, pad_token_label_id, mode=mode)
-    data_dir = Path(args.data_dir)
-    args.eval_batch_size = args.per_gpu_eval_batch_size * max(1, args.n_gpu)
-    eval_sampler = SequentialSampler(eval_dataset)
-    eval_dataloader = DataLoader(
-        eval_dataset,
-        sampler=eval_sampler,
-        batch_size=args.eval_batch_size,
-        collate_fn=None,
-    )
-
-    # Eval!
-    print("***** Running evaluation %s *****" % prefix)
-    print("  Num examples = %d" % len(eval_dataset))
-    print("  Batch size = %d" % args.eval_batch_size)
-    eval_loss = 0.0
-    nb_eval_steps = 0
-    preds = None
-    out_label_ids = None
-    model.eval()
-    files = []
-    perturb_prob = 0
-    mask_bbox_prob = 0.1
-    left_bbox_prob = 0.2
-    right_bbox_prob = 0.3
-    narrow_bbox_prob = 0.4
-    amplify_bbox_prob = 0.5
-
-    print(f" perturb_prob = {perturb_prob}")
-    print(f" mask_bbox_prob = {mask_bbox_prob}")
-    print(f" left_bbox_prob = {left_bbox_prob}")
-    print(f" right_bbox_prob = {right_bbox_prob}")
-    print(f" narrow_bbox_prob = {narrow_bbox_prob}")
-    print(f" amplify_bbox_prob = {amplify_bbox_prob}")
-
-    for batch in tqdm(eval_dataloader, desc="Evaluating"):
-        with torch.no_grad():
-            input_ids = batch[0]
-            new_ids = torch.randint_like(input_ids, tokenizer.vocab_size)
-            prob_mat = torch.rand(input_ids.shape)
-            input_ids = torch.where(prob_mat < perturb_prob, new_ids, input_ids)
-            inputs = {
-                "input_ids": input_ids.to(args.device),
-                "attention_mask": batch[1].to(args.device),
-                "labels": batch[3].to(args.device),
-            }
-            if args.model_type in ["layoutlm"]:
-                perturb_bbox = torch.clone(batch[4])
-                max_x = max(perturb_bbox[:, :, 0].max(), perturb_bbox[:, :, 2].max())
-                max_y = max(perturb_bbox[:, :, 1].max(), perturb_bbox[:, :, 3].max())
-                zero_bbox = torch.zeros_like(perturb_bbox)
-
-                B, L, D = perturb_bbox.shape
-                left_bbox = torch.cat((perturb_bbox[:, :-1, :], torch.zeros((B, 1, D), dtype=torch.int64)), dim=1)
-                right_bbox = torch.cat((perturb_bbox[:, 1:, :], torch.zeros((B, 1, D), dtype=torch.int64)), dim=1)
-
-                scale_bbox = torch.randint(0, 10, (B, L, 2))
-                narrow_bbox = torch.cat((scale_bbox, -scale_bbox), dim=2)
-                amplify_bbox = torch.cat((-scale_bbox, scale_bbox), dim=2)
-                prob_mat = torch.rand((B, L)).unsqueeze(dim=2)
-                prob_stack = prob_mat.repeat(1, 1, D)
-                # Mask perturb_bbox
-                perturb_bbox = torch.where(prob_stack < mask_bbox_prob, zero_bbox, perturb_bbox)
-                # Replace by left perturb_bbox
-                left_prob_mat = torch.where(prob_stack > left_bbox_prob, torch.zeros_like(prob_stack), prob_stack)
-                perturb_bbox = torch.where(mask_bbox_prob <= left_prob_mat, left_bbox, perturb_bbox)
-                # Replace by right perturb_bbox
-                right_prob_mat = torch.where(prob_stack > right_bbox_prob, torch.zeros_like(prob_stack), prob_stack)
-                perturb_bbox = torch.where(left_bbox_prob <= right_prob_mat, right_bbox, perturb_bbox)
-                # Narrow perturb_bbox
-                narrow_prob_mat = torch.where(prob_stack > narrow_bbox_prob, torch.zeros_like(prob_stack),
-                                              prob_stack)
-                perturb_bbox = torch.where(right_bbox_prob <= narrow_prob_mat, narrow_bbox, perturb_bbox)
-                # Amplify perturb_bbox
-                amplify_prob_mat = torch.where(prob_stack > amplify_bbox_prob, torch.zeros_like(prob_stack),
-                                               prob_stack)
-                perturb_bbox = torch.where(narrow_bbox_prob <= amplify_prob_mat, amplify_bbox, perturb_bbox)
-                perturb_bbox = torch.where(perturb_bbox < 0, zero_bbox, perturb_bbox)
-                ones_bbox = torch.ones((B, L), dtype=torch.int64)
-                # Replace items greater than max
-                perturb_bbox[:, :, 0] = torch.where(perturb_bbox[:, :, 0] > max_x, ones_bbox * max_x, perturb_bbox[:, :, 0])
-                perturb_bbox[:, :, 1] = torch.where(perturb_bbox[:, :, 1] > max_y,  ones_bbox * max_y, perturb_bbox[:, :, 1])
-                perturb_bbox[:, :, 2] = torch.where(perturb_bbox[:, :, 2] > max_x, ones_bbox * max_x, perturb_bbox[:, :, 2])
-                perturb_bbox[:, :, 3] = torch.where(perturb_bbox[:, :, 3] > max_y, ones_bbox * max_y, perturb_bbox[:, :, 3])
-                left_upper_points = perturb_bbox[:, :, :2]
-                right_bottom_points = perturb_bbox[:, :, 2:]
-                left_upper_bbox = torch.cat((left_upper_points, left_upper_points), dim=2)
-                right_bottom_bbox = torch.cat((right_bottom_points, right_bottom_points), dim=2)
-                perturb_bbox = torch.where(left_upper_bbox > right_bottom_bbox, left_upper_bbox, perturb_bbox)
-                assert torch.all(perturb_bbox[:, :, 2:] >= perturb_bbox[:, :, :2])
-                assert torch.all(perturb_bbox >= 0)
-                if amplify_bbox_prob > 0:
-                    inputs["bbox"] = perturb_bbox.to(args.device)
-                else:
-                    inputs["bbox"] = batch[4].to(args.device)
-            inputs["token_type_ids"] = (
-                batch[2].to(args.device)
-                if args.model_type in ["bert", "layoutlm"]
-                else None
-            )  # RoBERTa don"t use segment_ids
-            outputs = model(**inputs)
-            tmp_eval_loss, logits = outputs[:2]
-
-            if args.n_gpu > 1:
-                tmp_eval_loss = (
-                    tmp_eval_loss.mean()
-                )  # mean() to average on multi-gpu parallel evaluating
-            files += list(batch[5])
-            eval_loss += tmp_eval_loss.item()
-        nb_eval_steps += 1
-        if preds is None:
-            preds = logits.detach().cpu().numpy()
-            out_label_ids = inputs["labels"].detach().cpu().numpy()
-        else:
-            preds = np.append(preds, logits.detach().cpu().numpy(), axis=0)
-            out_label_ids = np.append(
-                out_label_ids, inputs["labels"].detach().cpu().numpy(), axis=0
-            )
-    # results = {"loss": eval_loss / nb_eval_steps}
-    probs, preds = torch.max(F.softmax(torch.tensor(preds), dim=2), dim=2)
-
-    label_map = {i: label for i, label in enumerate(labels)}
-
-    out_label_list = [[] for _ in range(out_label_ids.shape[0])]
-    preds_list = [[] for _ in range(out_label_ids.shape[0])]
-    preds_classes = [[] for _ in range(out_label_ids.shape[0])]
-    flatten_label_ids = out_label_ids.reshape((-1,)).tolist()
-    flatten_pred_ids = torch.flatten(preds).numpy().tolist()
-    flatten_label_list = []
-    flatten_pred_list = []
-    for label_id, pred_id in zip(flatten_label_ids, flatten_pred_ids):
-        if label_id != pad_token_label_id:
-            flatten_label_list.append(label_map[label_id])
-            flatten_pred_list.append(label_map[pred_id])
-    all_texts = eval_dataset.all_text
-    for i in range(out_label_ids.shape[0]):
-        tokens = []
-        for j in range(out_label_ids.shape[1]):
-            if out_label_ids[i, j] != pad_token_label_id:
-                token = tokenizer.convert_ids_to_tokens(eval_dataset.all_input_ids[i][j].item())
-                tokens.append(token)
-                out_label_list[i].append(label_map[out_label_ids[i][j]])
-                preds_list[i].append(preds[i][j].item())
-                preds_classes[i].append(label_map[preds[i][j].item()])
-    print("Data:" + data_dir.parent.stem)
-    annotation_path = data_dir.parent / 'test' / 'entities'
-    probs = probs.cpu().numpy().tolist()
-    num_pred = 0
-    num_correct = 0
-    num_gt = 0
-    if args.eval_strict:
-        print("****Strict eval model, which will compare entity****")
-        for file, text, pred_one_doc, probs in zip(files, all_texts, preds_list, probs):
-            pred_entity = convert_predictions_to_dict(label_map, text, pred_one_doc, probs)
-            num_pred += len(pred_entity)
-
-            with open(os.path.join(annotation_path, file + ".txt"), 'r') as f:
-                gt_entity = json.load(f)
-            num_gt += len(gt_entity)
-            for key, value in gt_entity.items():
-                if key in pred_entity and pred_entity[key] == value:
-                    num_correct += 1
-        precision = num_correct / num_pred
-        recall = num_correct / num_gt
-        results = {"precision": precision,
-                   "recall": recall,
-                   "f1": 2 * precision * recall / (precision + recall) if precision != 0 and recall != 0 else 0
-                   }
-    else:
-        results = {
-            "precision": precision_score(flatten_label_list, flatten_pred_list),
-            "recall": recall_score(flatten_label_list, flatten_pred_list),
-            "f1": f1_score(flatten_label_list, flatten_pred_list),
-        }
-    report = classification_report(flatten_label_list, flatten_pred_list)
-    print(report)
-    print("***** Eval results %s *****")
-    for key in sorted(results.keys()):
-        print(f"{key} = {100 * results[key]:.2f}%")
-
-    return results, preds_classes
-
 
 def main():  # noqa C901
-    parser = argparse.ArgumentParser()
-
-    ## Required parameters
-    parser.add_argument(
-        "--data_dir",
-        default=None,
-        type=str,
-        required=True,
-        help="The input data dir. Should contain the training files for the CoNLL-2003 NER task.",
-    )
-    parser.add_argument(
-        "--model_type",
-        default=None,
-        type=str,
-        required=True,
-        help="Model type selected in the list: " + ", ".join(MODEL_CLASSES.keys()),
-    )
-    parser.add_argument(
-        "--model_name_or_path",
-        default=None,
-        type=str,
-        required=True,
-        help="Path to pre-trained model or shortcut name selected in the list: "
-             + ", ".join(ALL_MODELS),
-    )
-    parser.add_argument(
-        "--output_dir",
-        default=None,
-        type=str,
-        required=True,
-        help="The output directory where the model predictions and checkpoints will be written.",
-    )
-
-    ## Other parameters
-    parser.add_argument(
-        "--labels",
-        default="",
-        type=str,
-        help="Path to a file containing all labels. If not specified, CoNLL-2003 labels are used.",
-    )
-    parser.add_argument(
-        "--config_name",
-        default="",
-        type=str,
-        help="Pretrained config name or path if not the same as model_name",
-    )
-    parser.add_argument(
-        "--tokenizer_name",
-        default="",
-        type=str,
-        help="Pretrained tokenizer name or path if not the same as model_name",
-    )
-    parser.add_argument(
-        "--cache_dir",
-        default="",
-        type=str,
-        help="Where do you want to store the pre-trained models downloaded from s3",
-    )
-    parser.add_argument(
-        "--max_seq_length",
-        default=512,
-        type=int,
-        help="The maximum total input sequence length after tokenization. Sequences longer "
-             "than this will be truncated, sequences shorter will be padded.",
-    )
-    parser.add_argument(
-        "--do_train", action="store_true", help="Whether to run training."
-    )
-    parser.add_argument(
-        "--do_eval", action="store_true", help="Whether to run eval on the dev set."
-    )
-    parser.add_argument(
-        "--do_predict",
-        action="store_true",
-        help="Whether to run predictions on the test set.",
-    )
-    parser.add_argument(
-        "--evaluate_during_training",
-        action="store_true",
-        help="Whether to run evaluation during training at each logging step.",
-    )
-    parser.add_argument(
-        "--do_lower_case",
-        action="store_true",
-        help="Set this flag if you are using an uncased model.",
-    )
-
-    parser.add_argument(
-        "--per_gpu_train_batch_size",
-        default=8,
-        type=int,
-        help="Batch size per GPU/CPU for training.",
-    )
-    parser.add_argument(
-        "--per_gpu_eval_batch_size",
-        default=8,
-        type=int,
-        help="Batch size per GPU/CPU for evaluation.",
-    )
-    parser.add_argument(
-        "--gradient_accumulation_steps",
-        type=int,
-        default=1,
-        help="Number of updates steps to accumulate before performing a backward/update pass.",
-    )
-    parser.add_argument(
-        "--learning_rate",
-        default=5e-5,
-        type=float,
-        help="The initial learning rate for Adam.",
-    )
-    parser.add_argument(
-        "--weight_decay", default=0.0, type=float, help="Weight decay if we apply some."
-    )
-    parser.add_argument(
-        "--adam_epsilon", default=1e-8, type=float, help="Epsilon for Adam optimizer."
-    )
-    parser.add_argument(
-        "--max_grad_norm", default=1.0, type=float, help="Max gradient norm."
-    )
-    parser.add_argument(
-        "--num_train_epochs",
-        default=3.0,
-        type=float,
-        help="Total number of training epochs to perform.",
-    )
-    parser.add_argument(
-        "--max_steps",
-        default=-1,
-        type=int,
-        help="If > 0: set total number of training steps to perform. Override num_train_epochs.",
-    )
-    parser.add_argument(
-        "--warmup_steps", default=0, type=int, help="Linear warmup over warmup_steps."
-    )
-
-    parser.add_argument(
-        "--logging_steps", type=int, default=50, help="Log every X updates steps."
-    )
-    parser.add_argument(
-        "--save_steps",
-        type=int,
-        default=50,
-        help="Save checkpoint every X updates steps.",
-    )
-    parser.add_argument(
-        "--eval_all_checkpoints",
-        action="store_true",
-        help="Evaluate all checkpoints starting with the same prefix as model_name ending and ending with step number",
-    )
-    parser.add_argument(
-        "--no_cuda", action="store_true", help="Avoid using CUDA when available"
-    )
-    parser.add_argument(
-        "--overwrite_output_dir",
-        action="store_true",
-        help="Overwrite the content of the output directory",
-    )
-    parser.add_argument(
-        "--overwrite_cache",
-        action="store_true",
-        help="Overwrite the cached training and evaluation sets",
-    )
-    parser.add_argument(
-        "--seed", type=int, default=42, help="random seed for initialization"
-    )
-
-    parser.add_argument(
-        "--fp16",
-        action="store_true",
-        help="Whether to use 16-bit (mixed) precision (through NVIDIA apex) instead of 32-bit",
-    )
-    parser.add_argument(
-        "--fp16_opt_level",
-        type=str,
-        default="O1",
-        help="For fp16: Apex AMP optimization level selected in ['O0', 'O1', 'O2', and 'O3']."
-             "See details at https://nvidia.github.io/apex/amp.html",
-    )
-    parser.add_argument(
-        "--local_rank",
-        type=int,
-        default=-1,
-        help="For distributed training: local_rank",
-    )
-    parser.add_argument(
-        "--eval_strict",
-        type=bool,
-        default=False,
-        help="For distributed training: local_rank",
-    )
-    parser.add_argument(
-        "--server_ip", type=str, default="", help="For distant debugging."
-    )
-    parser.add_argument(
-        "--server_port", type=str, default="", help="For distant debugging."
-    )
-    args = parser.parse_args()
+    parser = HfArgumentParser(TrainingArgs)
+    args = parser.parse_args_into_dataclasses()[0]
 
     if (
             os.path.exists(args.output_dir)
@@ -811,7 +179,7 @@ def main():  # noqa C901
         train_dataset = FunsdDataset(
             args, tokenizer, labels, pad_token_label_id, mode="train"
         )
-        global_step, tr_loss = train(
+        global_step, tr_loss = do_train(
             args, train_dataset, model, tokenizer, labels, pad_token_label_id
         )
         logger.info(" global_step = %s, average loss = %s", global_step, tr_loss)
@@ -835,85 +203,12 @@ def main():  # noqa C901
         torch.save(args, os.path.join(args.output_dir, "training_args.bin"))
 
     # Evaluation
-    results = {}
+    results = None
     if args.do_eval and args.local_rank in [-1, 0]:
-        tokenizer = tokenizer_class.from_pretrained(
-            args.output_dir, do_lower_case=args.do_lower_case
-        )
-        checkpoints = [args.output_dir]
-        if args.eval_all_checkpoints:
-            checkpoints = list(
-                os.path.dirname(c)
-                for c in sorted(
-                    glob.glob(args.output_dir + "/**/" + WEIGHTS_NAME, recursive=True)
-                )
-            )
-            logging.getLogger("pytorch_transformers.modeling_utils").setLevel(
-                logging.WARN
-            )  # Reduce logging
-        logger.info("Evaluate the following checkpoints: %s", checkpoints)
-        for checkpoint in checkpoints:
-            global_step = checkpoint.split("-")[-1] if len(checkpoints) > 1 else ""
-            model = model_class.from_pretrained(checkpoint)
-            model.to(args.device)
-            result, _ = evaluate(
-                args,
-                model,
-                tokenizer,
-                labels,
-                pad_token_label_id,
-                mode="test",
-                prefix=global_step,
-            )
-            if global_step:
-                result = {"{}_{}".format(global_step, k): v for k, v in result.items()}
-            results.update(result)
-        output_eval_file = os.path.join(args.output_dir, "eval_results.txt")
-        with open(output_eval_file, "w") as writer:
-            for key in sorted(results.keys()):
-                writer.write("{} = {}\n".format(key, str(results[key])))
+        results = do_eval(args, tokenizer_class, model_class, labels, pad_token_label_id)
 
     if args.do_predict and args.local_rank in [-1, 0]:
-        tokenizer = tokenizer_class.from_pretrained(
-            args.model_name_or_path, do_lower_case=args.do_lower_case
-        )
-        model = model_class.from_pretrained(args.output_dir)
-        model.to(args.device)
-        result, predictions = evaluate(
-            args, model, tokenizer, labels, pad_token_label_id, mode="test"
-        )
-        # Save results
-        output_test_results_file = os.path.join(args.output_dir, "test_results.txt")
-        with open(output_test_results_file, "w") as writer:
-            for key in sorted(result.keys()):
-                writer.write("{} = {}\n".format(key, str(result[key])))
-        # Save predictions
-        output_test_predictions_file = os.path.join(
-            args.output_dir, "test_predictions.txt"
-        )
-        with open(output_test_predictions_file, "w", encoding="utf8") as writer:
-            with open(
-                    os.path.join(args.data_dir, "test.txt"), "r", encoding="utf8"
-            ) as f:
-                example_id = 0
-                for line in f:
-                    if line.startswith("-DOCSTART-") or line == "" or line == "\n":
-                        writer.write(line)
-                        if not predictions[example_id]:
-                            example_id += 1
-                    elif predictions[example_id]:
-                        output_line = (
-                                line.split()[0]
-                                + " "
-                                + predictions[example_id].pop(0)
-                                + "\n"
-                        )
-                        writer.write(output_line)
-                    else:
-                        logger.warning(
-                            "Maximum sequence length exceeded: No prediction for '%s'.",
-                            line.split()[0],
-                        )
+        results = do_predict(args, tokenizer_class, model_class, labels, pad_token_label_id)
 
     return results
